Split babel into its own webpack config.
